+++ src/lexer/tokens.rs
@@ -10,15 +10,42 @@ pub enum TokenType {
 	Const, // const myvar = 1 (immutable)
 	Fn, // fn mycoolfn(arg: string) {}
     Return, // return mycoolvalue
+    If,
+    Else,
+    True,
+    False,
+
+    // TYPES FOR VARIABLES (reserved because static typing is a strong feature of cobalt)
+    // All types are 64-bit, 32-bit is a thing of the past
+    TypeInt,
+    TypeFloat,
+    TypeBool,
+    TypeStr,
 
     // Operators and Symbols
     Colon, // :  
-    BinaryOperator, // + - * / %
-    ComparisonOperator, // < > == <= >= !=
+    Plus, // +
+    Minus, // -
+    Star, // *
+    Slash, // /
+    Percent, // %
+    LessThan, // <
+    GreaterThan, // >
+    EqualsEquals, // ==
+    NotEqual, // !=
+    LessThanEqual, // <=
+    GreaterThanEqual, // >=
+    PlusEquals, // +=
+    MinusEquals, // -=
+    Or, // ||
+    And, // &&
     Equals, // =
+    Not, // ! (when found alone by itself, meant for unary expressions)
     OpenParen, // (
     CloseParen, // )
-    EOF // pushed at end of tokenization
+    OpenBrace, // {
+    CloseBrace, // }
+    EOF, // pushed at end of tokenization
 }
 
 #[derive(Debug, Clone)]