+++ src/parser/ast.rs
@@ -1,26 +1,36 @@
 #[derive(Debug, Clone, PartialEq)]
 pub enum NodeType {
-    Program,
-    VariableDeclaration,
-    BinaryExpr,
-    Identifier,
-    NumericLiteral,
-    AssignmentExpr,
-    UnaryExpr,
+    Program,             // the node that contains the AST
+    VariableDeclaration, // let x = 42, const x = 42 for immutable vars
+    IfStatement,         // if true {} else if x {} else {}, it checks the condition and if evaluated to true executes the statement following it.
+    BlockStatement,      // { body }, blocks have their own scope
+    BinaryExpr,          // an expression which has a left and right hand side seperated by an operator that determines the operation
+    LogicalExpr,         // an expression which has a left and right hand side seperated by an operator, either and (&&) or or (||).
+    Identifier,          // a name used to identify variables and functions declared by the user
+    NumericLiteral,      // 123
+    FloatLiteral,        // 123.0 (NOT IMPLEMENTED)
+    BooleanLiteral,      // true / false
+    StringLiteral,       // "content here" (NOT IMPLEMENTED)
+    AssignmentExpr,      // x = 42
+    UnaryExpr,           // -42, !true
 }
 
 #[derive(Debug, Clone)]
 pub enum Stmt {
     Program(Program),
     VariableDeclaration(VariableDeclaration),
+    IfStatement(IfStatement),
+    BlockStatement(BlockStatement),
     Expr(Expr),
 }
 
 #[derive(Debug, Clone)]
 pub enum Expr {
     Binary(BinaryExpr),
+    LogicalExpr(LogicalExpr),
     Identifier(Identifier),
     NumericLiteral(NumericLiteral),
+    BooleanLiteral(BooleanLiteral),
     AssignmentExpr(AssignmentExpr),
     UnaryExpr(UnaryExpr),
 }
@@ -42,11 +52,34 @@ pub struct VariableDeclaration {
 }
 
 #[derive(Debug, Clone)]
+pub struct IfStatement {
+    pub kind: NodeType,
+    pub test: Expr,
+    pub alternate: Option<Box<Stmt>>,
+    pub body: Box<Stmt> // either a single statement or a block statement are most common
+}
+
+// { code in here }
+#[derive(Debug, Clone)]
+pub struct BlockStatement {
+    pub kind: NodeType,
+    pub body: Vec<Stmt>,
+}
+
+#[derive(Debug, Clone)]
 pub struct BinaryExpr {
     pub kind: NodeType,
     pub left: Box<Expr>,
     pub right: Box<Expr>,
     pub operator: String,
+}
+
+#[derive(Debug, Clone)]
+pub struct LogicalExpr {
+    pub kind: NodeType,
+    pub left: Box<Expr>,
+    pub right: Box<Expr>,
+    pub operator: String
 }
 
 #[derive(Debug, Clone)]
@@ -62,9 +95,16 @@ pub struct NumericLiteral {
 }
 
 #[derive(Debug, Clone)]
+pub struct BooleanLiteral {
+    pub kind: NodeType,
+    pub value: bool,
+}
+
+#[derive(Debug, Clone)]
 pub struct AssignmentExpr {
     pub kind: NodeType,
     pub assignee: Box<Expr>,
+    pub operator: String, // =, +=, -=, /=, *=, %=
     pub value: Box<Expr>
 }